requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
colorama==0.4.6
Django==5.1.6
django-debug-toolbar==5.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
flake8==7.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
